Toplevel Mtags: use single indexer implementation that support all dialects
