+++ tests/unit/src/test/resources/workspace-symbol/example/nested/LocalClass.scala
@@ -2,6 +2,6 @@ package example.nested
 
 class LocalClass/*example.nested.LocalClass#*/ {
   def foo(): Unit = {
-    case class LocalClass/*example.nested.LocalClass#LocalClass#*/()
+    case class LocalClass()
   }
 }
